installation et config Dotenv + modif error 400 vers 500 + creation fichier .env et .env.example

+++ controllers/books.js
@@ -1,6 +1,29 @@
 const Book = require("../models/Book");
+const fs = require("fs");
 
-exports.addBook = (req, res, next) => {
+exports.getAllBooks = (req, res, next) => {
+  Book.find()
+    .then((books) => res.status(200).json(books))
+    .catch((error) => res.status(500).json({ error }));
+};
+
+exports.getOneBook = (req, res, next) => {
+  Book.findOne({ _id: req.params.id })
+    .then((book) => res.status(200).json(book))
+    .catch((error) => res.status(404).json({ error }));
+};
+
+exports.getBestBooks = (req, res, next) => {
+  Book.find()
+    .then((books) => {
+      books.sort((a, b) => b.averageRating - a.averageRating);
+      const bestRatedBooks = books.slice(0, 3);
+      res.status(200).json(bestRatedBooks);
+    })
+    .catch((error) => res.status(500).json({ error }));
+};
+
+exports.addOneBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book);
   delete bookObject._id;
   delete bookObject._userId;
@@ -15,9 +38,8 @@ exports.addBook = (req, res, next) => {
     .save()
     .then(() => res.status(201).json({ message: "Livre enregistré !" }))
     .catch((error) => {
-      res.status(400).json({ error });
+      res.status(500).json({ error });
     });
-  console.log(book);
 };
 
 exports.modifyBook = (req, res, next) => {
@@ -29,7 +51,6 @@ exports.modifyBook = (req, res, next) => {
         }`,
       }
     : { ...req.body };
-
   delete bookObject._userId;
   Book.findOne({ _id: req.params.id })
     .then((book) => {
@@ -42,27 +63,34 @@ exports.modifyBook = (req, res, next) => {
         )
           .then(() => res.status(200).json({ message: "Objet modifié !" }))
           .catch((error) => {
-            res.status(401).json({ error });
+            res.status(500).json({ error });
           });
       }
     })
-    .catch((error) => res.status(400).json({ error }));
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Livre supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
-};
-
-exports.getOneBook = (req, res, next) => {
   Book.findOne({ _id: req.params.id })
-    .then((book) => res.status(200).json(book))
-    .catch((error) => res.status(404).json({ error }));
+    .then((book) => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: "Non-authorisé" });
+      } else {
+        const filename = book.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Objet supprimé !" }))
+            .catch((error) => {
+              res.status(500).json({ error });
+            });
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
 
-exports.getAllBooks = (req, res, next) => {
-  Book.find()
-    .then((books) => res.status(200).json(books))
-    .catch((error) => res.status(400).json({ error }));
-};
+// exports.addRating = (req, res, next) => {
+
+// };
